requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
defusedxml==0.7.1
Django==6.0.3
django-allauth==65.15.1
django-cart==3.0.2
django-phonenumber-field==8.4.0
idna==3.11
oauthlib==3.3.1
phonenumbers==9.0.27
pillow==12.2.0
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
python3-openid==3.2.0
reportlab==4.4.10
requests==2.33.1
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
urllib3==2.2.1
